Weather App
This project is a small weather app implemented in Python. It uses the WeatherAPI to fetch real-time weather data for a specified city. The app allows users to view and hear the current temperature, weather condition, humidity, wind speed, and pressure for the selected city. It provides a menu-driven interface for easy interaction and utilizes the pyttsx3 library to provide text-to-speech functionality.