Image Resizer
Commit Information:

Commit Message: "Add image resizing functionality using OpenCV"

Commit Description:
This commit adds code that enables image resizing functionality using the OpenCV library in Python. The script allows users to resize images by specifying a scaling percentage. The source image is loaded using the cv2.imread() function from OpenCV, and the desired scaling percentage is applied to calculate the new dimensions. The image is then resized using the cv2.resize() function and saved to the specified destination file using cv2.imwrite(). The code also includes a print statement confirming the successful resizing of the image.

Modified Files:

image_resizer.py: Contains the implementation of image resizing functionality using OpenCV.
Dependencies:

OpenCV for Python: Required for image manipulation and resizing.
Usage:

Set the source variable to the path of the source image file.
Set the destination variable to the desired path for the resized image.
Set the scale_percent variable to the desired scaling percentage.
Execute the script to resize the image and save it to the specified destination.
Note: Uncomment the lines related to cv2.imshow(), cv2.waitKey(), and cv2.destroyAllWindows() if you want to display the resized image.

This code provides a convenient way to resize images using the OpenCV library, allowing users to customize the scaling percentage and save the resized image to a desired location.