FACE RECOGNITION ATTENDANCE  SYSTEM
The Face Recognition Attendance System is a Python project that uses face recognition technology to automate attendance tracking. It captures video from the default camera, detects faces, matches them against known faces, and records attendance for recognized individuals. The system is easy to use and generates a CSV file with the attendance records for the current date.